third_party/blink/web_tests/media/W3C/audio/currentSrc/currentSrc_empty_if_no_src-expected.txt

audio.currentSrc - empty if no source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "" is ""
PASS successfullyParsed is true

TEST COMPLETE


spec reference
